[refactor] moved apollo graphql related stuff to src/apollo
